Link user to meaning

+++ src/main/java/com/abbrevio/abbrevio/service/impl/MeaningServiceImpl.java
@@ -4,11 +4,15 @@ import com.abbrevio.abbrevio.dto.MeaningDTO;
 import com.abbrevio.abbrevio.entity.Abbreviation;
 import com.abbrevio.abbrevio.entity.Comment;
 import com.abbrevio.abbrevio.entity.Meaning;
+import com.abbrevio.abbrevio.entity.User;
 import com.abbrevio.abbrevio.exception.CustomNotFoundException;
 import com.abbrevio.abbrevio.repository.AbbreviationRepository;
 import com.abbrevio.abbrevio.repository.MeaningRepository;
+import com.abbrevio.abbrevio.repository.UserRepository;
 import com.abbrevio.abbrevio.service.MeaningService;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,12 +23,14 @@ public class MeaningServiceImpl implements MeaningService {
 
     private final MeaningRepository meaningRepository;
     private final AbbreviationRepository abbreviationRepository;
+    private final UserRepository userRepository;
     private final ModelMapper modelMapper;
 
-    public MeaningServiceImpl(MeaningRepository meaningRepository, ModelMapper modelMapper, AbbreviationRepository abbreviationRepository) {
+    public MeaningServiceImpl(MeaningRepository meaningRepository, ModelMapper modelMapper, AbbreviationRepository abbreviationRepository, UserRepository userRepository) {
         this.meaningRepository = meaningRepository;
         this.modelMapper = modelMapper;
         this.abbreviationRepository = abbreviationRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
@@ -51,12 +57,19 @@ public class MeaningServiceImpl implements MeaningService {
 
     @Override
     public MeaningDTO createMeaningForAbbreviation(MeaningDTO meaningDTO, Long id) {
+
         Abbreviation abbreviation = abbreviationRepository.findById(id)
                 .orElseThrow(() -> new CustomNotFoundException(Abbreviation.class, "id", id));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = userRepository.findByUsername(authentication.getName())
+                .orElseThrow(() -> new CustomNotFoundException(User.class, "username", authentication.getName()));
 
         Meaning meaning = new Meaning();
         meaning.setDescription(meaningDTO.getDescription());
         meaning.setAbbreviation(abbreviation);
+        meaning.setUser(user);
         meaning.setCountOfVotes((long) 0);
         meaningRepository.save(meaning);
 
@@ -64,12 +77,20 @@ public class MeaningServiceImpl implements MeaningService {
     }
 
     @Override
-    public MeaningDTO updateMeaningForAbbreviation(Long abbreviationId, Long meaningId, MeaningDTO meaningDTO){
+    public MeaningDTO updateMeaningForAbbreviation(Long abbreviationId, Long meaningId, MeaningDTO meaningDTO) throws Exception {
         Meaning meaning = meaningRepository.findByIdAndAbbreviationId(meaningId, abbreviationId);
 
         if (meaning == null)
         {
             throw new CustomNotFoundException(Meaning.class, "id", meaningId);
+        }
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userRepository.findByUsername(authentication.getName())
+                .orElseThrow(() -> new CustomNotFoundException(User.class, "username", authentication.getName()));
+
+        if (!meaning.getUser().getId().equals(user.getId()))
+        {
+            throw new Exception(String.format("you can't edit meaning with id %s, because you haven't created it", meaning.getId()));
         }
 
         if (meaningDTO.getDescription() != null) {
@@ -82,12 +103,19 @@ public class MeaningServiceImpl implements MeaningService {
     }
 
     @Override
-    public void deleteMeaningForAbbreviation(Long abbreviationId, Long meaningId){
+    public void deleteMeaningForAbbreviation(Long abbreviationId, Long meaningId) throws Exception {
         Meaning meaning = meaningRepository.findByIdAndAbbreviationId(meaningId, abbreviationId);
 
         if (meaning == null)
         {
             throw new CustomNotFoundException(Meaning.class, "id", meaningId);
+        }
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userRepository.findByUsername(authentication.getName())
+                .orElseThrow(() -> new CustomNotFoundException(User.class, "username", authentication.getName()));
+        if (!meaning.getUser().getId().equals(user.getId()))
+        {
+            throw new Exception(String.format("you can't edit meaning with id %s, because you haven't created it", meaning.getId()));
         }
 
         meaningRepository.delete(meaning);
